Handle timing issue bug in the client_handler
The client_handler can throw a KeyError when a STATUS_RES is received
after a WORK_COMPLETE for a given job.  The exact order of events looks
something like:

1) Server sends a WORK_COMPLETE
2) Client sends a GET_STATUS
3) Client receives the WORK_COMPLETE, and deletes the job_handle
4) Client receives a STATUS_RES, and tries to look up the job_handle.
   However, since it was deleted by the WORK_COMPLETE handler, a
   KeyError is thrown.

This workaround keeps old requests around in an old_handles dictionary.
TODO: make sure that really old handles are eventually deleted from this
dict.